+++ app/api/ada/restoration/byron/scan.js
@@ -22,7 +22,7 @@ import type {
 } from '../../lib/storage/database/walletTypes/common/utils';
 import type { AddByHashFunc, } from '../../lib/storage/bridge/hashMapper';
 import type { CanonicalAddressInsert } from '../../lib/storage/database/primitives/tables';
-import { CoreAddressTypes } from '../../lib/storage/database/primitives/tables';
+import { CoreAddressTypes } from '../../lib/storage/database/primitives/enums';
 import type { Bip44ChainInsert } from '../../lib/storage/database/walletTypes/common/tables';
 
 declare var CONFIG: ConfigType;
